+++ SAMLAssertion.properties
@@ -1,8 +1,4 @@
-#Below properties should be changed before executing the project.
 tokenUrl=
 clientId=
 userId=
-#userName will only be considered when userId is null/empty.
-userName=
 privateKey=
-expireInMinutes=
